First set of changes to bring unit test naming up to the standards in the style guide
More changes to unit tests to match conventions

Changed coding back to utf-8
